+++ src/Helpers/SpellHelper.php
@@ -182,6 +182,8 @@ class SpellHelper
             'wizard_strength' => '%+g%% wizard power',
             'wonder_damage' => 'Deals damage to wonders',
             'scale_by_day' => 'Scales by day in round from 137.5%% to 62.5%%',
+            'explore_cost_wizard_mastery' => 'Exploring platinum cost reduced by 1%% per %d Wizard Mastery',
+            'spell_refund' => 'Failed and reflected spells refund %d%% of their mana cost',
         ];
     }