+++ src/Services/Dominion/Actions/InvadeActionService.php
@@ -987,10 +987,14 @@ class InvadeActionService
         }
 
         // Special case for Ascendance
-        if (isset($survivingUnits[1]) && $dominion->getSpellPerkValue('upgrade_swordsmen') && $this->invasionResult['result']['range'] >= 75) {
-            $ascendedUnits = floor($survivingUnits[1] * $dominion->getSpellPerkValue('upgrade_swordsmen') / 100);
-            $convertedUnits[1] -= $ascendedUnits;
-            $convertedUnits[4] += $ascendedUnits;
+        foreach ($dominion->race->units as $unit) {
+            $perkValue = $dominion->race->getUnitPerkValueForUnitSlot($unit->slot, 'upgrade_survivors');
+            if (!$perkValue || !array_key_exists($unit->slot, $survivingUnits) || ($survivingUnits[$unit->slot] === 0)) {
+                continue;
+            }
+            $upgradedUnits = floor($survivingUnits[$unit->slot] * $perkValue[1] / 100);
+            $convertedUnits[$unit->slot] -= $upgradedUnits;
+            $convertedUnits[$perkValue[0]] += $upgradedUnits;
         }
 
         if (!isset($this->invasionResult['attacker']['conversion']) && $convertedUnits !== array_fill(1, 4, 0)) {